requirements.txt
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
google-cloud-core==2.4.1
html2text==2024.2.26
